# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/leads": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all leads grouped by stage for the Kanban pipeline view.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Leads"
                ],
                "summary": "Get Kanban board",
                "responses": {
                    "200": {
                        "description": "Map of stage → []Lead",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new lead and broadcasts a lead.created WebSocket event.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Leads"
                ],
                "summary": "Create lead",
                "parameters": [
                    {
                        "description": "Lead payload (contact_id required)",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_dynamicweblab_masaar-crm_internal_domain.Lead"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_dynamicweblab_masaar-crm_internal_domain.Lead"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/leads/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a single lead with its contact joined.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Leads"
                ],
                "summary": "Get lead",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Lead UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_dynamicweblab_masaar-crm_internal_domain.Lead"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/leads/{id}/stage": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates the pipeline stage of a lead (drag-drop). Broadcasts lead.stage_changed event.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Leads"
                ],
                "summary": "Move lead to stage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Lead UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New stage: new|contacted|qualified|proposal|won|lost",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "stage": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "lead_id": {
                                    "type": "string"
                                },
                                "stage": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_dynamicweblab_masaar-crm_internal_domain.Contact": {
            "type": "object",
            "properties": {
                "assigned_to": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "language": {
                    "description": "\"ar\" | \"en\"",
                    "type": "string"
                },
                "lead_score": {
                    "type": "integer"
                },
                "phone_wa": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "github_com_dynamicweblab_masaar-crm_internal_domain.Lead": {
            "type": "object",
            "properties": {
                "contact": {
                    "description": "Joined",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_dynamicweblab_masaar-crm_internal_domain.Contact"
                        }
                    ]
                },
                "contact_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "currency": {
                    "description": "default: AED",
                    "type": "string"
                },
                "deal_value": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "source": {
                    "$ref": "#/definitions/github_com_dynamicweblab_masaar-crm_internal_domain.LeadSource"
                },
                "stage": {
                    "$ref": "#/definitions/github_com_dynamicweblab_masaar-crm_internal_domain.LeadStage"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_dynamicweblab_masaar-crm_internal_domain.LeadSource": {
            "type": "string",
            "enum": [
                "whatsapp",
                "web",
                "referral",
                "event"
            ],
            "x-enum-varnames": [
                "SourceWhatsApp",
                "SourceWeb",
                "SourceReferral",
                "SourceEvent"
            ]
        },
        "github_com_dynamicweblab_masaar-crm_internal_domain.LeadStage": {
            "type": "string",
            "enum": [
                "new",
                "contacted",
                "qualified",
                "proposal",
                "won",
                "lost"
            ],
            "x-enum-varnames": [
                "StageNew",
                "StageContacted",
                "StageQualified",
                "StageProposal",
                "StageWon",
                "StageLost"
            ]
        },
...
    }
}